Update system for generating PDFS
Add ability to customize the latex generated by pandoc so as to
include various bits of metadata that are required in regulatory
documents.

This approach is somewhat fragile, since it is editing the text using
pattern matching, and because it depends on the form of the output that
pandoc creates.

We can improve this system over time, but I think it will work well for
the foreseeable future.