Move SOUP details into SDS
There was no good reason to have them in a separate document, except
that that is how I have seen it done before.

It seems to make sense to include it in the SDS with the descriptions of
the other software items.

Also, the soup components document was an odd mixture of process and
documentation; this new split feels better since the process stuff is in
the software plan and the documentation stuff is in the SDS.