Document and simplify YAML front matter
- Remove the unused "Category" key
- Make the "revision" key optional, since there will typically not be
revisions for records (records are a special type of document)
- Move the "manufacturer" key into `system.yml`, since it is likely to
be the same for all documents. I considered pulling from the frontmatter
if present, and falling back to `system.yml`, but it seemed like
over-engineering at this point. The best argument to do this is so that
the tex is generated from ONLY the release markdown files.
- Add some documentation for the YAML front matter.