Make default makefile's clean recipe ONLY remove PDFs
It can be convenient to delete document/XXX.md but not delete
release/XXX.md. This lets you keep a release copy without re-making it
over and over. This pattern is esp useful for records.

The problem is, the default make file deletes the entire release
directoyry.

Change this behaviour so ONLY pdfs are deleted by default.