Make it possible to specifiy custom classes for the backend
Also update and simplify the documentation a bit. I removed some documentation
of parts of the API which may change, e.g. how we configure markdown
extensions, since I think in a future release we will want to change
this. In particular, I think we will need a mechanism to pass
configuration to extensions (maybe even on a per-template basis).
Hopefully we can find a generic way to do this so that each extension
can work the same way... I suspect it will somehow have to do with YAML
front matter overriding defaults in config.yml.