Update image path related settings
When compiling PDFs and DOCX files, download the media to the
`tmp/media` folder.

Otherwise, use paths relative from the Make file. Note that this will
break image display when the release markdown files are viewed in
GitHub. In a future release we'll reccomend using GitHub pages to view
these outputs, instead of committing the release documents to the repo
itself.