Don't include 14971 in the 62304 checklists
Often the risk management plan is handled at the system level and is
outside of the purveiew of RDM. Thus, most of the 14971 checklist items
aren't covered by the software plan. Rather, we rely on 62304:4.2 to
pull in the 14971 compliant plan and leave the checklist off. If a user
wants to verify their risk managent plan against 14971 they can run that
audit separately.

If the 14971 checklist is included in the 62304 checklist, there's no
easy way to exclude it if it isn't desired (as is often the case).