Fill in a basic software description document using the 2021 guidance
The new 2021 draft FDA guidance, which will likely replace the 2005
guidance, includes more specific details regarding what should be
included in the software description document. The 2005 document just
had a short paragraph followed by this statement:

> If this information is included in another document, such as the
> Software Requirements Specification, your submission should contain an
> annotation and a reference to the document in the submission where this
> information is located.

Thus, we used to just include a placeholder document with a reference to
the SDS. Now that the FDA has provided more details regarding what
should go in this document, we've decided to include it, directly
following the questions that the FDA followed in the guidance document.
We believe following the format that the FDA expects will make it easier
on reviewers at the FDA.

This template includes several pre-filled sections and is thus slightly
opinionated.

We still expect that many of the questions will be answered with
references.